artif: add .aliases and .*_aliases to shell artif
Signed-off-by: Adam V. Link <linkavych@gmail.com>

This commit adds a '.aliases' and '._*aliases'
search pattern to the files/shell/config.yaml
artifact. Linux and Unix users will commonly
place aliases in a separate file, and is sourced
to their main configuration file (e.g. .bashrc).
Attacker can abuse these aliases, and these
should be collected for analysis.

NOTE: if the shell configuration file is a symlink
in the directory, the collector will not identify
the file for collection.