move tool logs to the front of archive
This moves the tool's logs to the front of the tar archive rather than
the tail. As tar is a stream format, having them at the tail ends up
requiring that automation read the archive twice: once to gather details
crucial to routing and parsing the data it contains, and once to actually
parse the data.

Most solutions will likely still want to read archives twice in order to
handle ones generated before this change, but at least they can
terminate the first pass when they finish parsing the log file, thus
minimizing waste.